first user handling added first user checking is added files moved
